# Reuse optimized IR for bytecode dependencies