--- a/Changelog.md
+++ b/Changelog.md
@@ -6,6 +6,7 @@ Language Features:
 Compiler Features:
  * EVM: Support for the EVM version "Prague".
  * SMTChecker: Add CHC engine check for underflow and overflow in unary minus operation.
+ * Yul Optimizer: Reuse optimized Yul IR of contract dependencies instead of optimizing them again along with the dependent contract.
 
 
 Bugfixes: