The quick and easy solution for #15179, which only reuses the IR.