@cameel 2024-06-05
Benchmarks on external projects:

| Project              | v0.8.26   | This PR  | Speedup |
|----------------------|----------:|---------:|--------:|
| openzeppelin         |    40 s   |     36 s | 10%     |
| uniswap-v4           |   177 s   |    116 s | 34%     |