@cameel 2024-06-07
Here's some more benchmarking with #15183:

| Project              | legacy | IR (0.8.26) | IR (this PR) | Speedup   |
|----------------------|-------:|------------:|-------------:|----------:|
| openzeppelin         |   10 s |        41 s |         36 s |       12% |
| uniswap-v4           |   22 s |       259 s |        140 s |       46% |
| eigenlayer           |   80 s |      1324 s |        622 s |       53% |
| sablier-v2           |  153 s |      1952 s |        889 s |       54% |

Some important notes:
- This time the numbers are averages from several runs. The small difference for OZ is actually reproducible and not just a fluke.
- The previous results for Uniswap were on a commit from 2024-05-10. I updated to 2024-06-06 before this benchmark and apparently it now compiles *a lot* slower even with a release binary. I think they must have added more tests. Still, this seems to make the effect of the PR even more pronounced.
- Sablier fails with StackTooDeep when compiled via IR. I still included it because this seems to happen towards the end compilation and the result seems in line with the other projects in the table, all of which did complete without errors.