@cameel 2024-06-07
And here is the timing from external tests (extracted using the script from https://github.com/ethereum/solidity/pull/15023#issuecomment-2067151960):

| Project          |   Before |   After |   Diff |
|:-----------------|---------:|--------:|-------:|
| brink            |      4 s |     4 s |    0 s |
| colony           |    100 s |    99 s |   -1 s |
| elementfi        |     98 s |    82 s |  -16 s |
| ens              |     26 s |    35 s |    9 s |
| euler            |     30 s |    32 s |    2 s |
| gnosis           |          |         |        |
| gp2              |     38 s |    30 s |   -8 s |
| pool-together    |     35 s |    33 s |   -2 s |
| uniswap          |     47 s |    37 s |  -10 s |
| yield_liquidator |     22 s |    22 s |    0 s |
| zeppelin         |    177 s |   149 s |  -28 s |

They numbers don't seem particularly reliable though. For example the longer running time for `ens` is unexpected and I suspect it's just the CI variance, but then most of the other results don't tell us anything either because they're in a similar range.

Also, I'm not sure how OpenZeppelin in CI takes ~150 s to compile. The main difference is that it's being compiled with Hardhat rather than Foundry, but such a difference still seems odd. Is it downloading something rather than just building? Is the CI machine that slow?