@cameel 2024-07-16
I'm closing this since we won't be merging it. The actual fix is based on caching optimized IR instead.